agents-sdk/templates/python/requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
solana>=0.36.0
solders>=0.25.0
